requirements.txt
streamlit 
sentence-transformers 
faiss-cpu 
openai 
tiktoken 
pydantic 
pandas 
numpy 
scikit-learn
